relocate world_size()
