+++ torch_xla/_internal/pjrt.py
@@ -66,7 +66,7 @@ def _run_thread_per_device(
     torch_xla._XLAC._xla_set_default_device(device)
 
     # See Note Note [Dynamo WORLD_SIEZ and ORDINAL].
-    xm._init_world_size_ordinal()
+    runtime._init_world_size_ordinal()
 
     return fn()
